Moce saveLog function into respective actions instead of component doing it
